Move find node to test utils
